Move CTAPHID commands under a single vendor command
Previously, we used one CTAPHID vendor command per application command.
This makes it hard to extend the functionality because we have to
synchronize the used vendor commands over all applications in the
Trussed ecosystem.

As an alternative, this patch introduces a new vendor command that acts
as a namespace for the admin application.  The actual application
command is encoded in the first byte of the payload.  In the future, we
might want to make this configurable so that the runner can decide the
namespace used by CTAPHID apps.

For compatibility, the CTAPHID vendor commands are still available.

This change also allows us to combine the CTAPHID and APDU command
handling into a single dispatch function.